# backend/openapi.yaml
openapi: 3.2.0
info:
  title: Shuuen Backend API
  version: 1.0.0
  license:
    name: The Supreme Mandate of the Void 1.2
    url: https://github.com/username-ldv/shuuen-backend/blob/main/LICENSE
  description: >-
    Public catalog reads and role-aware account access. Catalog mutations require
    an administrator bearer token. Private reads additionally require
    include_private=true.
servers:
  - url: /
    description: Current origin
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  schemas:
...
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, pattern: '^[A-Za-z0-9_]{3,20}$' }
        password: { type: string, minLength: 8, maxLength: 72 }
        display_name: { type: string, maxLength: 160 }
    AuthResult:
      type: object
      required: [user, access_token, token_type, expires_at]
      properties:
        user: { $ref: '#/components/schemas/User' }
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role or private scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register an ordinary user when registration is enabled
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRequest' }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/AuthResult' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Username already registered }
        '429': { description: Rate limit exceeded }
  /api/v1/auth/login:
    post:
      summary: Log in
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/AuthResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { description: Rate limit exceeded }
  /api/v1/auth/me:
    get:
      summary: Current user
      operationId: getCurrentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/password:
    post:
      summary: Change password and revoke previously issued tokens
      operationId: changePassword
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, maxLength: 72 }
                new_password: { type: string, minLength: 8, maxLength: 72 }
      responses:
        '200':
          description: Password changed and replacement token issued
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/courses:
    get:
      summary: List public courses without level payloads
      operationId: listCourses
      security: [{}, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/IncludePrivate' }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Course summary page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoursePage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create an empty managed course and its top-level library folder
      operationId: createCourse
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 180 }
                slug: { type: string, maxLength: 220 }
                description: { type: string }
                author: { type: string, maxLength: 180 }
                is_public: { type: boolean, default: false }
                sort_order: { type: integer, default: 0 }
      responses:
        '201': { description: Created course }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Course slug or folder already exists }
  /api/v1/courses/{courseId}:
    get:
      summary: Get course metadata, modes, progression groups, and counts without levels
      operationId: getCourse
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Course metadata and navigation hierarchy }
        '404': { description: Course not found or private }
    put:
      summary: Replace only course metadata; nested structure is untouched
      operationId: updateCourse
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, is_public]
              properties:
                name: { type: string, maxLength: 180 }
                description: { type: string }
                author: { type: string, maxLength: 180 }
                is_public: { type: boolean }
                sort_order: { type: integer }
      responses:
        '200': { description: Updated metadata }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Course not found }
  /api/v1/courses/{courseId}/{mode}:
    get:
      summary: Get one mode and its progression-group metadata and counts
      operationId: getCourseMode
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Course mode }
        '404': { description: Course or mode not found }
    put:
      summary: Replace one mode's metadata without touching its groups or levels
      operationId: updateCourseMode
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 180 }
                description: { type: string }
      responses:
        '200': { description: Updated mode }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Course or mode not found }
  /api/v1/courses/{courseId}/modes:
    post:
      summary: Add one mode to a course
      description: A folder blueprint is transparently made editable before the mode is added.
      operationId: createCourseMode
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: { type: string, enum: [singles, melodies, chords] }
                name: { type: string, maxLength: 180 }
                description: { type: string }
                position: { type: integer, minimum: 0 }
      responses:
        '201': { description: Created mode }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Mode or folder already exists }
  /api/v1/courses/{courseId}/{mode}/position:
    put:
      summary: Move a mode to a zero-based position using a lightweight payload
      operationId: positionCourseMode
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties: { position: { type: integer, minimum: 0 } }
      responses:
        '204': { description: Reordered }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/courses/{courseId}/{mode}/groups:
    post:
      summary: Add one independent progression group/tab
      operationId: createProgressionGroup
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 180 }
                slug: { type: string, maxLength: 220 }
                description: { type: string }
                position: { type: integer, minimum: 0 }
      responses:
        '201': { description: Created progression group }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Group folder already exists }
  /api/v1/courses/{courseId}/{mode}/groups/{groupId}:
    put:
      summary: Replace one progression group's metadata
      operationId: updateProgressionGroup
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 180 }
                description: { type: string }
      responses:
        '200': { description: Updated progression group }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: 'Course, mode, or group not found' }
  /api/v1/courses/{courseId}/{mode}/groups/{groupId}/position:
    put:
      summary: Move a progression group to a zero-based tab position
      operationId: positionProgressionGroup
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: groupId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties: { position: { type: integer, minimum: 0 } }
      responses:
        '204': { description: Reordered }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/courses/{courseId}/{mode}/levels:
    get:
      summary: Page through level payloads for a mode or progression group
      operationId: listCourseLevels
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: group_id
          in: query
          description: Stable progression group id returned by the metadata endpoint.
          schema: { type: string }
        - name: ids
          in: query
          description: Comma-separated level ids, at most 200; response preserves requested order.
          schema: { type: string }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200':
          description: Level page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseLevelPage' }
        '404': { description: 'Course, mode, or group not found' }
    post:
      summary: Add one level; the rest of the course is untouched
      operationId: createCourseLevel
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group_id, name, source, definition]
              properties:
                id: { type: string, maxLength: 64 }
                group_id: { type: string }
                name: { type: string, maxLength: 220 }
                source: { type: string, enum: [built_in, user, imported] }
                definition: { $ref: '#/components/schemas/CourseLevelDefinition' }
                is_public: { type: boolean, default: true }
                position: { type: integer, minimum: 0 }
      responses:
        '201': { description: Created level }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Level id already exists }
  /api/v1/courses/{courseId}/{mode}/levels/query:
    post:
      summary: Fetch up to 200 levels by id without URL-length pressure
      operationId: queryCourseLevels
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - { $ref: '#/components/parameters/IncludePrivate' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  minItems: 1
                  maxItems: 200
                  uniqueItems: true
                  items: { type: string }
      responses:
        '200': { description: Existing visible levels in requested id order }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/courses/{courseId}/{mode}/levels/{levelId}:
    get:
      summary: Get one visible course level
      operationId: getCourseLevel
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: levelId
          in: path
          required: true
          schema: { type: string }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Course level }
        '404': { description: Level not found or private }
    put:
      summary: Replace one level's metadata and definition without changing its position
      operationId: updateCourseLevel
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: levelId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, source, definition, is_public]
              properties:
                name: { type: string, maxLength: 220 }
                source: { type: string, enum: [built_in, user, imported] }
                definition: { $ref: '#/components/schemas/CourseLevelDefinition' }
                is_public: { type: boolean }
      responses:
        '200': { description: Updated level }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Level not found }
    delete:
      summary: Remove one course-level record without deleting referenced MIDI files
      operationId: deleteCourseLevel
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: levelId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted level record; files preserved }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Level not found }
  /api/v1/courses/{courseId}/{mode}/levels/{levelId}/position:
    put:
      summary: Reorder a level or move it to another progression group
      operationId: positionCourseLevel
      security: [{ bearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: integer }
        - name: mode
          in: path
          required: true
          schema: { type: string, enum: [singles, melodies, chords] }
        - name: levelId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties:
                group_id:
                  type: string
                  description: Omit to reorder within the current progression group.
                position: { type: integer, minimum: 0 }
      responses:
        '204': { description: Reordered or moved }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: 'Course, mode, level, or target group not found' }
  /api/v1/library/groups:
    get:
      summary: List groups
      operationId: listGroups
      security: [{}, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/IncludePrivate' }
        - name: parent_id
          in: query
          schema: { type: integer }
        - name: parent_path
          in: query
          schema: { type: string }
        - name: path_prefix
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Group page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/groups/{id}:
    get:
      summary: Get a group and a bounded melody page
      operationId: getGroup
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/IncludePrivate' }
        - name: recursive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200': { description: Group tree response }
        '404': { description: Group not found or private }
...
  /api/v1/library/melodies:
    get:
      summary: List melodies
      operationId: listMelodies
      security: [{}, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/IncludePrivate' }
        - name: group_id
          in: query
          schema: { type: integer }
        - name: group_path
          in: query
          schema: { type: string }
        - name: recursive
          in: query
          schema: { type: boolean }
        - name: tag
          in: query
          schema: { type: string }
        - name: format
          in: query
          schema: { type: string, enum: [midi, musicxml] }
        - name: q
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Melody page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MelodyPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/tags:
    get:
      summary: List tags
      operationId: listTags
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Tag page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TagPage' }
    post:
      summary: Create a tag
      operationId: createTag
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                slug: { type: string, maxLength: 100 }
                color: { type: string, maxLength: 24 }
      responses:
        '201': { description: Created tag }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Name or slug already exists }
  /api/v1/library/tags/{id}:
    get:
      summary: Get a tag
      operationId: getTag
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Tag }
        '404': { description: Tag not found }
    patch:
      summary: Update a tag
      operationId: updateTag
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                slug: { type: string, maxLength: 100 }
                color: { type: string, maxLength: 24 }
      responses:
        '200': { description: Updated tag }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Name or slug already exists }
    delete:
      summary: Delete a tag
      operationId: deleteTag
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/melodies/{id}:
    get:
      summary: Get a melody
      operationId: getMelody
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Melody }
        '404': { description: Melody not found or private }
    delete:
      summary: Delete a melody and its files
      operationId: deleteMelody
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/melodies/{id}/variants:
    get:
      summary: List variants for a visible melody
      operationId: listVariants
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Variants }
    post:
      summary: Upload a variant
      operationId: uploadVariant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                format: { type: string, enum: [midi, musicxml] }
                file: { type: string, format: binary }
      responses:
        '201': { description: Variant created }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/variants/{id}:
    get:
      summary: Get variant metadata
      operationId: getVariant
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200': { description: Variant }
        '404': { description: Variant not found or private }
    patch:
      summary: Change primary status
      operationId: updateVariant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { is_primary: { type: boolean } }
      responses:
        '200': { description: Updated variant }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      summary: Delete a variant and its file
      operationId: deleteVariant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/library/variants/{id}/download:
    get:
      summary: Download a visible variant
      operationId: downloadVariant
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - { $ref: '#/components/parameters/IncludePrivate' }
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { description: Variant not found or private }
...